third_party/WebKit/LayoutTests/http/tests/sendbeacon/beacon-cookie-expected.txt

CONSOLE WARNING: line 18: Synchronous XMLHttpRequest on the main thread is deprecated because of its detrimental effects to the end user's experience. For more help, check https://xhr.spec.whatwg.org/.
PingLoader dispatched to 'http://127.0.0.1:8000/sendbeacon/resources/save-beacon.php?name=cookie'.
Checking transmission of Beacons involving cookies.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS navigator.sendBeacon("resources/save-beacon.php?name=cookie", "Blip"); is true
PASS Beacon sent successfully
PASS Content-Type: text/plain;charset=UTF-8
PASS Cookie: hello=world
PASS Origin: http://127.0.0.1:8000
PASS Referer: http://127.0.0.1:8000/sendbeacon/beacon-cookie.html
PASS Request-Method: POST
PASS Length: 4
PASS Body: Blip
PASS 
PASS successfullyParsed is true

TEST COMPLETE
